Move price modules to build/prices/ subpackage
Relocate price_builder, price_archive, price_writers, price_s3, and
price_fetcher into mtgjson5/build/prices/ for cleaner organization.
Update all import paths and doc references.
